+++ App.js
@@ -1,21 +1,10 @@
-import React from 'react';
-import MapView from 'react-native-maps';
-import { StyleSheet, View } from 'react-native';
+import React, { useEffect } from 'react';
+import AppNavigator from './lib/navigation/AppNavigator';
+import { LogBox } from 'react-native';
+LogBox.ignoreAllLogs();
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <MapView style={styles.map} />
-    </View>
+    <AppNavigator />
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  map: {
-    width: '100%',
-    height: '100%',
-  },
-});
