--- App_old.js
@@ -1,10 +1,0 @@
-// import React, { useEffect } from 'react';
-// import AppNavigator from './lib/navigation/AppNavigator';
-// import { LogBox } from 'react-native';
-// LogBox.ignoreAllLogs();
-
-// export default function App() {
-//   return (
-//     <AppNavigator />
-//   );
-// }
