+++ lib/screens/mapPage/googleMap.tsx
@@ -1,21 +1,21 @@
-// import React from 'react';
-// import MapView from 'react-native-maps';
-// import { StyleSheet, View } from 'react-native';
+import React from 'react';
+import MapView from 'react-native-maps';
+import { StyleSheet, View } from 'react-native';
 
-// export default function App() {
-//   return (
-//     <View style={styles.container}>
-//       <MapView style={styles.map} />
-//     </View>
-//   );
-// }
+export default function GoogleMap() {
+  return (
+    <View style={styles.container}>
+      <MapView style={styles.map} />
+    </View>
+  );
+}
 
-// const styles = StyleSheet.create({
-//   container: {
-//     flex: 1,
-//   },
-//   map: {
-//     width: '100%',
-//     height: '100%',
-//   },
-// });
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+  map: {
+    width: '100%',
+    height: '100%',
+  },
+});
